+++ adapter/persistence/src/main/java/de/strasser/peter/hexagonal/persistence/repository/CustomerRepository.java
@@ -4,6 +4,8 @@ import de.strasser.peter.hexagonal.persistence.model.CustomerEntity;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface CustomerRepository extends MongoRepository<CustomerEntity, Integer> {
+public interface CustomerRepository extends MongoRepository<CustomerEntity, BigInteger> {
 }
