+++ application/src/main/java/de/strasser/peter/hexagonal/application/customer/port/in/AddAddressUseCase.java
@@ -5,8 +5,9 @@ import de.strasser.peter.hexagonal.application.customer.port.in.commands.AddAddr
 import javax.validation.Valid;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotEmpty;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface AddAddressUseCase {
-    void addAddresses(@Min(0) Integer customerId, @Valid @NotEmpty List<AddAddressCommand> addresses);
+    void addAddresses(@Min(0) BigInteger customerId, @Valid @NotEmpty List<AddAddressCommand> addresses);
 }
